SDK bugfixes and improvements, sample app
1. DRY versioning used in User Agent, VERSION constant moved to
PublisherSDK class.
2. restricted url access to test-graph and graph on goplay.com.
3. __get needs to append to segment when it's the first argument to
allow eg. $sdk->pub->playersclub->delete(accountId);
4. SDKException is broadened to include the server error type and url
associated with the request in order to enable better reporting.

A sample app has also been added in sample.php to do the following:
1. show current accounts registered with test-graph.goplay.com.
2. show the current token (note: session is required for persistence of
the token).
3. allow a user to post data to the test-graph endpoint.
4. allow a user to delete an account by specifying accountId.

There is at present no error handling or validation in this app as it is
meant to be instructional or proof of concept only.